% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cal_viewer_fct_helpers.R
\name{import_cal}
\alias{import_cal}
\title{Import calendar directly from server}
\usage{
import_cal(cal_slug = "wimpys-world-of-streamers", cal_base_url = NULL)
}
\arguments{
\item{cal_slug}{string for URL slug of calendar.}
}
\value{
data frame with calendar event contents
}
\description{
Import calendar directly from server
}
